moving test and data files to be able to use/test metar without installation:
- move /reference/nsd_cccc.txt to /metar/data
- move metar/tests to /tests
- move test to /tests/data
- change file access in tests
- use pkg_resources to access nsd_cccc.txt as package data
