+++ metar/station.py
@@ -10,6 +10,8 @@ import os
 import sys
 import pdb
 import codecs
+from contextlib import closing
+from pkg_resources import resource_stream
 
 # math stuff
 import numpy as np
@@ -735,10 +737,9 @@ def _process_sky_cover(obs):
 
 
 def getAllStations():
-    stationfile = os.path.join(sys.prefix, 'metar_data', 'reference', 'nsd_cccc.txt')
     stations = {}
 
-    with open(stationfile, 'r') as fh:
+    with closing(resource_stream('metar.data', 'nsd_cccc.txt')) as fh:
         for line in fh:
             f = line.strip().split(";")
             stations[f[0]] = (f[0], f[3], f[4], f[5], f[7], f[8])
