+++ tests/data/test_dumpNCDCFormat.dat
@@ -1,0 +1,2 @@
+HPD0404168500HPCPHI20130100010200600 00541  0700 00392  0800 00639  0900 00410  1000 00563  1100 00560  1200 00594  1300 00718  1400 00600  1500 00397  1600 00669  1700 00335  1800 00316  1900 00359  2000 00753  2100 00704  2200 00319  2300 00705  2400 00329  2500 09903  
+HPD0404168500HPCPHI20130100020110100 00692  0200 00186  0300 00470  0400 00429  0500 00592  0600 00351  0700 00463  0800 00436  0900 00460  1000 00102  2500 04181
